+++ RubiksCube/CubeScramble.cpp
@@ -4,8 +4,6 @@
 #include <unordered_map>
 #include <functional>
 
-
-
 CubeScramble::CubeScramble() {
 
 }
