+++ RubiksCube/RubiksCube/x64/Debug/RubiksCube.tlog/Cl.items.tlog
@@ -1,4 +1,5 @@
 D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\CubeScramble.cpp;D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\RubiksCube\x64\Debug\CubeScramble.obj
+D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\Load2D.cpp;D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\RubiksCube\x64\Debug\Load2D.obj
 D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\RubiksCube.cpp;D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\RubiksCube\x64\Debug\RubiksCube.obj
 D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\TextRender.cpp;D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\RubiksCube\x64\Debug\TextRender.obj
 D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\Timer.cpp;D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\RubiksCube\x64\Debug\Timer.obj
