+++ RubiksCube/RubiksCube/x64/Debug/RubiksCube.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^D:\4.GODINA\RG\PROJEKAT\TEXTRENDERING23NOV\RUBIKS-CUBE-OPENGL\RUBIKSCUBE\RUBIKSCUBE\X64\DEBUG\CUBESCRAMBLE.OBJ|D:\4.GODINA\RG\PROJEKAT\TEXTRENDERING23NOV\RUBIKS-CUBE-OPENGL\RUBIKSCUBE\RUBIKSCUBE\X64\DEBUG\RUBIKSCUBE.OBJ|D:\4.GODINA\RG\PROJEKAT\TEXTRENDERING23NOV\RUBIKS-CUBE-OPENGL\RUBIKSCUBE\RUBIKSCUBE\X64\DEBUG\TEXTRENDER.OBJ|D:\4.GODINA\RG\PROJEKAT\TEXTRENDERING23NOV\RUBIKS-CUBE-OPENGL\RUBIKSCUBE\RUBIKSCUBE\X64\DEBUG\TIMER.OBJ
+^D:\4.GODINA\RG\PROJEKAT\TEXTRENDERING23NOV\RUBIKS-CUBE-OPENGL\RUBIKSCUBE\RUBIKSCUBE\X64\DEBUG\CUBESCRAMBLE.OBJ|D:\4.GODINA\RG\PROJEKAT\TEXTRENDERING23NOV\RUBIKS-CUBE-OPENGL\RUBIKSCUBE\RUBIKSCUBE\X64\DEBUG\LOAD2D.OBJ|D:\4.GODINA\RG\PROJEKAT\TEXTRENDERING23NOV\RUBIKS-CUBE-OPENGL\RUBIKSCUBE\RUBIKSCUBE\X64\DEBUG\RUBIKSCUBE.OBJ|D:\4.GODINA\RG\PROJEKAT\TEXTRENDERING23NOV\RUBIKS-CUBE-OPENGL\RUBIKSCUBE\RUBIKSCUBE\X64\DEBUG\TEXTRENDER.OBJ|D:\4.GODINA\RG\PROJEKAT\TEXTRENDERING23NOV\RUBIKS-CUBE-OPENGL\RUBIKSCUBE\RUBIKSCUBE\X64\DEBUG\TIMER.OBJ
 D:\4.godina\RG\Projekat\TextRendering23nov\rubiks-cube-opengl\RubiksCube\RubiksCube\x64\Debug\RubiksCube.ilk
